requirements.txt
python_telegram_bot
pillow